+++ types/protocol.d.ts
@@ -16895,6 +16895,46 @@ export namespace Protocol {
             devices: PromptDevice[];
         }
     }
+
+    export namespace Preload {
+
+        /**
+         * Unique id
+         */
+        export type RuleSetId = string;
+
+        /**
+         * Corresponds to SpeculationRuleSet
+         */
+        export interface RuleSet {
+            id: RuleSetId;
+            /**
+             * Identifies a document which the rule set is associated with.
+             */
+            loaderId: Network.LoaderId;
+            /**
+             * Source text of JSON representing the rule set. If it comes from
+             * <script> tag, it is the textContent of the node. Note that it is
+             * a JSON for valid case.
+             * 
+             * See also:
+             * - https://wicg.github.io/nav-speculation/speculation-rules.html
+             * - https://github.com/WICG/nav-speculation/blob/main/triggers.md
+             */
+            sourceText: string;
+        }
+
+        /**
+         * Upsert. Currently, it is only emitted when a rule set added.
+         */
+        export interface RuleSetUpdatedEvent {
+            ruleSet: RuleSet;
+        }
+
+        export interface RuleSetRemovedEvent {
+            id: RuleSetId;
+        }
+    }
 }
 
 export default Protocol;